refactor(listener): replace register file with listener directory
